+++ src/layers.py
@@ -1,6 +1,6 @@
 """Custom layers that are needed for the various models."""
 import tensorflow as tf
-from tensorflow import keras
+import tf_keras as keras
 
 
 class LinearRegression(keras.layers.Layer):
@@ -32,7 +32,10 @@ class LinearRegression(keras.layers.Layer):
             trainable=True,
             name='offset')
 
-    def call(self, inputs):
-        """Actually perform the calculation."""
+    def call(self, inputs, *args, **kwargs):
+        """Actually perform the calculation.
+
+        Variadic arguments are ignored.
+        """
         return inputs * self.slope + self.offset
 # Copyright 2022, 2023, 2024 Charles McAnany. This file is part of BPReveal. BPReveal is free software: You can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 2 of the License, or (at your option) any later version. BPReveal is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with BPReveal. If not, see <https://www.gnu.org/licenses/>.  # noqa