+++ src/trainSoloModel.py
@@ -152,7 +152,7 @@ import bpreveal.internal.disableTensorflowLogging  # pylint: disable=unused-impo
 from bpreveal import utils
 if __name__ == "__main__":
     utils.setMemoryGrowth()
-from tensorflow import keras  # pylint: disable=wrong-import-order
+import tf_keras as keras
 from bpreveal import logUtils
 from bpreveal import models
 import bpreveal.training